[Fix] 아이콘 상태 변화 기능 파일 구분
